requirements.txt
PySide6~=6.6.0
Pillow~=10.1.0
appdirs~=1.4.4
opencv-python~=4.9.0.80
numpy~=1.26.2
matplotlib~=3.8.2
mmdet~=3.3.0
pycocotools~=2.0.7